+++ src/usbc_average_lookup/__init__.py
@@ -1,4 +1,4 @@
 """USBC Average Lookup application foundation."""
 
-__version__ = "0.7.0a3"
+__version__ = "0.7.0a4"